backend/requirements.txt
Flask
twelvelabs==1.0.0
requests
python-dotenv
flask-cors
apscheduler
gunicorn
moviepy==1.0.3
Pillow>=10.2.0
numpy>=1.26.0
imageio-ffmpeg==0.4.9